Use hash rocket in the first DSL-like method call
In the rule for writing a DSL-like method call in multiple lines,
the first and second example were the same.

I assume hash rocket is used originally in the first example, but
when updating hash rocket to JSON style (i.e. `{ foo: 42 }`) in
the documents, it was turned into JSON style by mistake.