Update Python style guide
The current style guide in here is outdated. The search team in collaboration with ML and MLI (the main users of Python in Cookpad) decided to create a new one from scratch, matching the current state. The style guide has been stable for the last couple of month and these team consider it as the de facto style guide for Python.

The Python style guide is in a separate repository to allow and encourage fast iteration, ownership and collaboration of the teams that use it. So we believe it should stay that way. There are couple of options of what to do with the `cookpad/styleguide` for Python. I suggest linking to the `cookpad/python-style-guide`. However note that while `cookpad/styleguide` is public, `cookpad/python-style-guide` is private. So the link would work only for Cookpad employees. If that is a problem, we can either make `cookpad/python-style-guide` public or remove the link from here.

CC: @mattjw, @ettiee, @PermanentPon